requirements.txt
discord.py
requests
beautifulSoup4
schedule